refactor: crawler and stat -> root/crawler.c, root/stat.c
