+++ Makefile.am
@@ -71,11 +71,13 @@ watchman_SOURCES = \
 	main.c       \
 	root.c       \
 	root/ageout.c       \
+	root/crawler.c       \
 	root/dir.c       \
 	root/file.c       \
 	root/iothread.c       \
 	root/lock.c       \
 	root/notifythread.c       \
+	root/stat.c       \
 	root/symlink.c       \
 	state.c      \
 	string.c     \
